chore: split budget database by responsibility
